api/requirements.txt
flask
pony
PyMySQL
pycryptodomex